WareHouse/masterprompter_DefaultOrganization_20250202145847/requirements.txt
pyaudio==0.2.13
librosa==0.10.0.post2
soundfile==0.12.1
openai==0.28.1
openai-whisper
pyttsx3==2.90
